# [k8s] GPU Image debloat