Optimizes the Kubernetes GPU image for >2x faster provisioning (87s vs 201s). The new image is 1.96G compressed vs 6.39G compressed for the previous image.

```
# Master
sky launch -y -c test --cloud kubernetes --gpus T4:1 -- nvidia-smi  7.72s user 4.54s system 6% cpu 3:21.66 total

# This image
sky launch -y -c test2 --cloud kubernetes --gpus T4:1 -- nvidia-smi  4.74s user 4.30s system 9% cpu 1:27.87 total
```

Also adds kubectl to the dependencies pre-installed in the image for faster controller setup.

Closes #3435 and #2461.

Related - https://github.com/skypilot-org/skypilot-catalog/pull/72/

Tested:
- [x] Kubernetes smoke tests (in progress)