@romilbhardwaj 2024-06-14
Updated CUDA image to 12.1.1 since that will be supported till Feb 2025, and is at parity with cu121 offered on GCP. 

Running smoke tests again.