@romilbhardwaj 2024-06-18
Thanks @Michaelvll! Will merge after https://github.com/skypilot-org/skypilot-catalog/pull/72 is approved and merged.